requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.2.3
frozenlist==1.3.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gspread==5.8.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.15
yarl==1.9.2
